requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
black==21.5b1
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
click==8.0.0
idna==2.10
lxml==4.6.3
mypy-extensions==0.4.3
numpy==1.20.3
pandas==1.2.4
pathspec==0.8.1
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
urllib3==1.26.4
